+++ backend/app/api/v1/endpoints/users.py
@@ -1,53 +1,38 @@
 from typing import Any, List
 
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter
+from crud import crud_users
+from fastapi.encoders import jsonable_encoder
+
 
 router = APIRouter()
 
 
 @router.get("/")
-def get_users():
-    try:
-        return {}
-    except HTTPException:
-        pass 
+def get_users(): 
+    crud_users.CRUDUser.get_all
 
 
 @router.get("/{id}")
 def get_user_by_id(id):
-    try:
-        return {}
-    except HTTPException:
-        pass 
+    crud_users.CRUDUser.get_by_id(id)
 
 
 @router.get("/me")
 def get_my_user():
-    try:
-        return {}
-    except HTTPException:
-        pass 
+    crud_users.CRUDUser.get_my
 
 
 @router.post("/")
 def create_user():
-    try:
-        return {}
-    except HTTPException:
-        pass 
+    crud_users.CRUDUser.create
 
 
 @router.patch("/{id}")
-def modify_user():
-    try:
-        return {}
-    except HTTPException:
-        pass 
+def modify_user(id):
+    crud_users.CRUDUser.modify(id)
 
 
 @router.delete("/{id}")
 def delete_user(id):
-    try:
-        return {}
-    except HTTPException:
-        pass 
+    crud_users.CRUDUser.delete(id)